Remove Images, Updated Dashboard service, update schedule put calls

+++ Api.Core/UseCases/Generic/RemoveUseCase.cs
@@ -6,6 +6,8 @@ using Quiplogs.Core.Interfaces;
 using Quiplogs.Core.Interfaces.Repositories;
 using Quiplogs.Core.Interfaces.UseCases.Generic;
 using System.Threading.Tasks;
+using Quiplogs.BlobStorage;
+using Quiplogs.BlobStorage.Models;
 
 namespace Quiplogs.Core.UseCases.Generic
 {
@@ -14,18 +16,41 @@ namespace Quiplogs.Core.UseCases.Generic
         where T2 : BaseEntityDto
     {
         private readonly IBaseRepository<T1, T2> _repository;
+        private readonly IBlobStorage _blobStorage;
 
-        public RemoveUseCase(IBaseRepository<T1, T2> repository)
+        public RemoveUseCase(IBaseRepository<T1, T2> repository, IBlobStorage blobStorage)
         {
             _repository = repository;
+            _blobStorage = blobStorage;
         }
 
         public async Task<bool> Handle(RemoveRequest request, IOutputPort<RemoveResponse> outputPort)
         {
-            await _repository.Remove(request.Id);
+            var model = await _repository.GetById(request.Id);
 
-            outputPort.Handle(new RemoveResponse("Removed Successfully", true));
-            return true;
+            if (model.Success)
+            {
+                var imageModel = (IImage) model;
+                if (!string.IsNullOrWhiteSpace(imageModel.ImageFileName))
+                {
+                    var deleteImage = new DeleteFileRequest
+                    {
+                        Container = imageModel.CompanyId.ToString(),
+                        SubContainer = imageModel.Id.ToString(),
+                        FileName = imageModel.ImageFileName
+                    };
+
+                    await _blobStorage.DeleteBlobImage(deleteImage);
+                }
+                
+                await _repository.Remove(request.Id);
+
+                outputPort.Handle(new RemoveResponse("Removed Successfully", true));
+                return true;
+            }
+
+            outputPort.Handle(new RemoveResponse(model.Errors));
+            return false;
         }
     }
 }
